+++ pkg/kubelet/dockershim/docker_stats_test.go
@@ -1,4 +1,4 @@
-// +build !dockerless
+// +build linux,!dockerless
 
 /*
 Copyright 2019 The Kubernetes Authors.
